+++ app/models/subject.rb
@@ -48,15 +48,13 @@ class Subject < ActiveRecord::Base
     current_user.all_subjects.where(id: id).count == 1
   end
 
-  def self.first_or_create_with_defaults(project, subject_code, acrostic, user, default_site)
+  def self.first_or_create_with_defaults(project, subject_code, acrostic, site_name, user, default_site)
     # (1) Find existing subject...
     subject = project.subjects.where(subject_code: subject_code).first
     return subject if subject
     # (2) if not found slot into site by subject code and set proper site or use fallback
-    site = project.sites.find_by_id(project.site_id_with_prefix(subject_code))
-    if site
-      default_site = site
-    end
+    site = project.sites.find_by_name(site_name)
+    default_site = site if site
 
     subject = project.subjects.where(subject_code: subject_code).first_or_create(acrostic: acrostic, user_id: user.id, site_id: default_site.id)
     subject