+++ db/migrate/20160210162701_remove_columns_from_sites.rb
@@ -1,0 +1,7 @@
+class RemoveColumnsFromSites < ActiveRecord::Migration
+  def change
+    remove_column :sites, :prefix, :string, null: false, default: ''
+    remove_column :sites, :code_minimum, :string
+    remove_column :sites, :code_maximum, :string
+  end
+end